Reduce Warnings

+++ shared/basetsd.h
@@ -319,7 +319,7 @@ PtrToPtr32(
     const void *p
     )
 {
-    return((void * POINTER_32) (unsigned long) (ULONG_PTR) p);
+    return((void * POINTER_32) (ULONG_PTR) p);
 }
 
 #define HandleToHandle32( h )       (PtrToPtr32( h ))
